+++ lib/providers/user_provider.dart
@@ -1,10 +1,15 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../remote/firestore_planner_api.dart';
 import 'auth_provider.dart';
 
 final firestoreProvider = Provider<FirebaseFirestore>((ref) {
   return FirebaseFirestore.instance;
+});
+
+final firestorePlannerApiProvider = Provider<FirestorePlannerApi>((ref) {
+  return FirestorePlannerApi();
 });
 
 final currentUserIdProvider = Provider<String>((ref) {